+++ packages/web/package.json
@@ -22,13 +22,15 @@
     "react-dom": "18.2.0",
     "react-i18next": "^12.3.1",
     "turndown": "^7.1.2",
-    "lexical":"0.12.6",
+    "lexical": "0.12.6",
     "@lexical/react": "0.12.6",
+    "papaparse": "^5.4.1",
     "@lexical/utils": "0.12.6",
     "@lexical/text": "0.12.6"
   },
   "devDependencies": {
     "@types/react": "18.2.0",
+    "@types/papaparse": "^5.3.7",
     "@types/react-dom": "18.2.0",
     "@types/turndown": "^5.0.4"
   }
