+++ projects/app/package.json
@@ -42,7 +42,6 @@
     "next": "13.5.2",
     "next-i18next": "^13.3.0",
     "nprogress": "^0.2.0",
-    "papaparse": "^5.4.1",
     "react": "18.2.0",
     "react-day-picker": "^8.7.1",
     "react-dom": "18.2.0",
@@ -66,7 +65,6 @@
     "@types/jsonwebtoken": "^9.0.3",
     "@types/lodash": "^4.14.191",
     "@types/node": "^20.8.5",
-    "@types/papaparse": "^5.3.7",
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.0",
     "@types/react-syntax-highlighter": "^15.5.6",
