+++ projects/app/src/web/common/file/utils.ts
@@ -1,29 +1,3 @@
-import Papa from 'papaparse';
-import { readFileRawText } from '@fastgpt/web/common/file/read/rawText';
-
-/**
- * read csv to json
- * @response {
- *  header: string[],
- *  data: string[][]
- * }
- */
-export const readCsvContent = async (file: File) => {
-  try {
-    const { rawText: textArr } = await readFileRawText(file);
-    const csvArr = Papa.parse(textArr).data as string[][];
-    if (csvArr.length === 0) {
-      throw new Error('csv 解析失败');
-    }
-    return {
-      header: csvArr.shift() as string[],
-      data: csvArr.map((item) => item)
-    };
-  } catch (error) {
-    return Promise.reject('解析 csv 文件失败');
-  }
-};
-
 /**
  * file download by text
  */
